evaluation/requirements.txt
requests==2.32.3
google-genai==1.52.0
python-dotenv==1.2.1